Extract secret from gcs
update

move

update

update
